Feat: [EVER-129] 인증 기반 쿠폰 사용 API 구현

+++ src/main/java/com/team4ever/backend/domain/coupon/repository/UserCouponRepository.java
@@ -11,9 +11,10 @@ public interface UserCouponRepository extends JpaRepository<UserCoupon, Integer>
     // 사용자+쿠폰 단건 조회 (사용 상태 확인)
     @EntityGraph(attributePaths = "coupon")
     Optional<UserCoupon> findByUserIdAndCouponId(Long userId, Integer couponId);
+
     boolean existsByUserIdAndCouponId(Long userId, Integer couponId);
 
     // 사용자별 전체 쿠폰 조회 (isUsed 상태 확인)
     @EntityGraph(attributePaths = "coupon")
-    List<UserCoupon> findByUserId(Integer userId);
+    List<UserCoupon> findByUserId(Long userId);
 }
